Resolve issue https://github.com/boyter/lc/issues/9

+++ parsers/helpers.go
@@ -80,7 +80,7 @@ func licenceListHasLicense(license LicenseMatch, licenseList []LicenseMatch) boo
 // Apache-2.0 is compatible with GPL-3.0-only
 // See https://apache.org/licenses/GPL-compatibility.html
 // SPDX-License-Identifier: Apache-2.0
-func bytesToHuman(bytes int) string {
+func bytesToHuman(bytes int64) string {
 
 	const TERABYTE = 1099511627776
 	const GIGABYTE = 1073741824
